+++ RELEASES.md
@@ -22,6 +22,7 @@
 - registered types in the reflection system
 - added `InputMap::clear`
 - fixed [a bug](https://github.com/Leafwing-Studios/leafwing-input-manager/issues/430) related to incorrect axis data in `Chord` when not all buttons are pressed.
+- fixed [a bug](https://github.com/Leafwing-Studios/leafwing-input-manager/issues/446) related to missing restoration for `ButtonState::Pressed` state when re-enabling `ToggleActions`.
 
 ### Code Quality